+++ R/zzz.R
@@ -1,2 +1,8 @@
-globalVariables(c("lon", "lat", "value", "clustering", "output", "DD", "DR", "FF", "FH", "FX", "HH", "RH", "T10", "TD", "YYYYMMDD", "knmi_stations", "Y"))
+globalVariables(c("lon", "lat", "value", "clustering", "output", "DD", "DR",
+                  "FF", "FH", "FX", "HH", "RH", "T10", "TD",
+                  "YYYYMMDD", "knmi_stations", "Y",
+                  ".", "geohash_sum", "geohash", "longitude", "latitude",
+                  "delta_longitude", "delta_latitude", "gh_nghbrs_sum",
+                  "sum_bbox", "_sum_insured", "head", "valueconc",
+                  ":=", "north", "south", "west", "east", "geometry"))